Remove plugin architecture.
- Clean up legacy types.
