+++ test/unit/common/util.test.ts
@@ -110,71 +110,6 @@ describe("util", () => {
     })
   })
 
-  describe("getOptions", () => {
-    beforeEach(() => {
-      const location: LocationLike = {
-        pathname: "/healthz",
-        origin: "http://localhost:8080",
-        // search: "?environmentId=600e0187-0909d8a00cb0a394720d4dce",
-      }
-
-      // Because resolveBase is not a pure function
-      // and relies on the global location to be set
-      // we set it before all the tests
-      // and tell TS that our location should be looked at
-      // as Location (even though it's missing some properties)
-      global.location = location as Location
-    })
-
-    afterEach(() => {
-      jest.restoreAllMocks()
-    })
-
-    it("should return options with base and cssStaticBase even if it doesn't exist", () => {
-      expect(util.getClientConfiguration()).toStrictEqual({
-        base: "",
-        csStaticBase: "",
-      })
-    })
-
-    it("should return options when they do exist", () => {
-      // Mock getElementById
-      const spy = jest.spyOn(document, "getElementById")
-      // Create a fake element and set the attribute
-      const mockElement = document.createElement("div")
-      mockElement.setAttribute(
-        "data-settings",
-        '{"base":".","csStaticBase":"./static/development/Users/jp/Dev/code-server","logLevel":2,"disableUpdateCheck":false}',
-      )
-      // Return mockElement from the spy
-      // this way, when we call "getElementById"
-      // it returns the element
-      spy.mockImplementation(() => mockElement)
-
-      expect(util.getClientConfiguration()).toStrictEqual({
-        base: "",
-        csStaticBase: "/static/development/Users/jp/Dev/code-server",
-        disableUpdateCheck: false,
-        logLevel: 2,
-      })
-    })
-
-    it("should include queryOpts", () => {
-      // Trying to understand how the implementation works
-      // 1. It grabs the search params from location.search (i.e. ?)
-      // 2. it then grabs the "options" param if it exists
-      // 3. then it creates a new options object
-      // spreads the original options
-      // then parses the queryOpts
-      location.search = '?options={"logLevel":2}'
-      expect(util.getClientConfiguration()).toStrictEqual({
-        base: "",
-        csStaticBase: "",
-        logLevel: 2,
-      })
-    })
-  })
-
   describe("arrayify", () => {
     it("should return value it's already an array", () => {
       expect(util.arrayify(["hello", "world"])).toStrictEqual(["hello", "world"])
