--- test/unit/node/test-plugin/.gitignore
@@ -1,1 +1,0 @@
-out
