--- test/unit/node/test-plugin/package.json
@@ -1,16 +1,0 @@
-{
-  "private": true,
-  "name": "test-plugin",
-  "version": "1.0.0",
-  "engines": {
-    "code-server": "^3.7.0"
-  },
-  "main": "out/index.js",
-  "devDependencies": {
-    "@types/express": "^4.17.8",
-    "typescript": "^4.0.5"
-  },
-  "scripts": {
-    "build": "tsc"
-  }
-}
