--- test/unit/node/test-plugin/public/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Test Plugin</title>
-  </head>
-  <body>
-    <p>Welcome to the test plugin!</p>
-  </body>
-</html>
